requirements.txt
alembic==1.4.1
aniso8601==9.0.1
astroid==2.6.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
cryptography==3.4.7
cycler==0.10.0
databricks-cli==0.14.3
docker==5.0.0
entrypoints==0.3
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
gitdb==4.0.7
GitPython==3.1.18
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
isort==5.9.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
mlflow==1.18.0
numpy==1.21.0
packaging==21.0
pandas==1.3.0
Pillow==8.3.0
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
protobuf==3.17.3
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
smmap==4.0.0
SQLAlchemy==1.4.20
sqlparse==0.4.1
tabulate==0.8.9
threadpoolctl==2.1.0
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
urllib3==1.26.6
websocket-client==1.1.0
Werkzeug==2.0.1
wrapt==1.12.1
